+++ tests/test_mnist.py
@@ -9,6 +9,8 @@ from phgrad.nn import MLP
 from phgrad.optim import SGD
 from phgrad.loss import nllloss
 
+from phgrad import types
+
 
 @pytest.mark.slow
 class TestMNIST(unittest.TestCase):
@@ -29,7 +31,7 @@ class TestMNIST(unittest.TestCase):
             for i in range(0, len(self.X_train), 32):
                 optimizer.zero_grad()
                 x = Tensor(self.X_train[i : i + 32])
-                y = Tensor(np.argmax(self.Y_train[i : i + 32], axis=1))
+                y = Tensor(np.argmax(self.Y_train[i : i + 32], axis=1), dtype=types.int64)
                 y_pred = mlp(x)
                 y_pred_log_softmax = y_pred.log_softmax(dim=1)
                 loss = nllloss(y_pred_log_softmax, y, reduce="mean")
@@ -110,7 +112,7 @@ class TestMNISTCUDA(unittest.TestCase):
             for i in range(0, len(self.X_train), 32):
                 optimizer.zero_grad()
                 x = Tensor(self.X_train[i : i + 32], device="cuda")
-                y = Tensor(np.argmax(self.Y_train[i : i + 32], axis=1), device="cuda")
+                y = Tensor(np.argmax(self.Y_train[i : i + 32], axis=1), device="cuda", dtype=types.int64)
                 y_pred = mlp(x)
                 y_pred_log_softmax = y_pred.log_softmax(dim=1)
                 y_pred_log_softmax = y_pred.log_softmax(dim=1)
